correccion del modulo
