+++ Controller/Index/Index.php
@@ -1,0 +1,29 @@
+<?php
+namespace Bootcamp\Repoback\Controller\Index;
+
+
+class Index implements \Magento\Framework\App\Action\HttpGetActionInterface
+{
+    /**
+     * @var \Magento\Framework\View\Result\PageFactory
+     */
+    protected $_pageFactory;
+
+
+    public function __construct(
+       
+       \Magento\Framework\View\Result\PageFactory $pageFactory
+    )
+    {
+        $this->_pageFactory = $pageFactory;
+    }
+    /**
+     * View page action
+     *
+     * @return \Magento\Framework\Controller\ResultInterface
+     */
+    public function execute()
+    {
+        return $this->_pageFactory->create();
+    }
+}
