--- HolaMundoK/composer.json
@@ -1,17 +1,0 @@
-{
-    "name": "bootcampkc/holamundok",
-    "description": "Este modulo saluda al mundo",
-    "type": "magento2-module",
-    "version": "1.0.0", 
-    "license": [
-      "MIT"
-    ],
-    "autoload": {
-      "psr-4": {
-        "Bootcampkc\\HolaMundoK\\": ""
-      },
-      "files": [
-        "registration.php"
-      ]
-    }
-  }
