--- HolaMundoK/i18n/es_CO.csv
@@ -1,2 +1,0 @@
-"Index Hello World", "Index Hola Mundo"
-"Hello World PRO", "Hola Mundo PRO"
