--- HolaMundoK/view/frontend/layout/holamundokaren_index_index.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0"?>
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="1column" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <body>
-        <referenceContainer name="content">
-            <block  name="bootcampkc.index" template="Bootcampkc_HolaMundoK::index.phtml" />
-        </referenceContainer>
-    </body>
-</page>
