--- HolaMundoK/view/frontend/templates/index.phtml
@@ -1,5 +1,0 @@
-<?php
-/** @var \Magento\Framework\View\Element\Template $block */
-?>
-
-<h1><?= __("Index Hello World") ?></h1>
