--- HolaMundoW/etc/frontend/routes.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:App/etc/routes.xsd">
-    <router id="standard">
-        <route id="helloworld" frontName="holamundoW">
-            <module name="Repoback_HolaMundoW" />
-        </route>
-    </router>
-</config>
