--- HolaMundoW/view/frontend/templates/index.phtml
@@ -1,6 +1,0 @@
-<?php
-/** @var \Repoback\HolaMundoW\Block\Template $block */
-?>
-<div class="template">
-    <h1><?=__('HELLO WORLD WILBERT')?></h1>
-</div>
