--- PageBuilderBlockTheme/Setup/Patch/Data/html/amazon_slider_main.html
@@ -1,144 +1,0 @@
-<style>
-    #html-body [data-pb-style=XUAMDU7] {
-        justify-content: flex-start;
-        display: flex;
-        flex-direction: column;
-        background-position: left top;
-        background-size: contain;
-        background-repeat: no-repeat;
-        background-attachment: scroll
-    }
-
-    #html-body [data-pb-style=BEW6KGQ] {
-        text-align: center;
-        min-height: 300px;
-        margin: 0;
-        padding: 0
-    }
-
-    #html-body [data-pb-style=W5VIMR7] {
-        background-position: left top;
-        background-size: cover;
-        background-repeat: no-repeat;
-        min-height: 300px
-    }
-
-    #html-body [data-pb-style=TEBEEB2] {
-        min-height: 300px;
-        background-color: transparent
-    }
-
-    #html-body [data-pb-style=X5PYI8J] {
-        background-position: left top;
-        background-size: cover;
-        background-repeat: no-repeat;
-        min-height: 300px
-    }
-
-    #html-body [data-pb-style=GL4B38B] {
-        min-height: 300px;
-        background-color: transparent
-    }
-
-    #html-body [data-pb-style=G1KKEN9] {
-        background-position: left top;
-        background-size: cover;
-        background-repeat: no-repeat;
-        min-height: 300px
-    }
-
-    #html-body [data-pb-style=WXHSFDH] {
-        min-height: 300px;
-        background-color: transparent
-    }
-
-    #html-body [data-pb-style=JMQW2EB] {
-        background-position: left top;
-        background-size: cover;
-        background-repeat: no-repeat;
-        min-height: 300px
-    }
-
-    #html-body [data-pb-style=SSS7R2N] {
-        min-height: 300px;
-        background-color: transparent
-    }
-</style>
-<div data-content-type="row" data-appearance="contained" data-element="main">
-    <div data-enable-parallax="0" data-parallax-speed="0.5" data-background-images="{}" data-background-type="image"
-        data-video-loop="true" data-video-play-only-visible="true" data-video-lazy-load="true"
-        data-video-fallback-src="" data-element="inner" data-pb-style="XUAMDU7">
-        <div class="pagebuilder-slider" data-content-type="slider" data-appearance="default" data-autoplay="true"
-            data-autoplay-speed="4000" data-fade="false" data-infinite-loop="true" data-show-arrows="false"
-            data-show-dots="false" data-element="main" data-pb-style="BEW6KGQ">
-            <div data-content-type="slide" data-slide-name="" data-appearance="poster" data-show-button="never"
-                data-show-overlay="never" data-element="main">
-                <div data-element="empty_link">
-                    <div class="pagebuilder-slide-wrapper"
-                        data-background-images="{\&quot;desktop_image\&quot;:\&quot;{{media url=wysiwyg/4.jpg}}\&quot;,\&quot;mobile_image\&quot;:\&quot;{{media url=wysiwyg/4a.jpg}}\&quot;}"
-                        data-background-type="image" data-video-loop="true" data-video-play-only-visible="true"
-                        data-video-lazy-load="true" data-video-fallback-src="" data-element="wrapper"
-                        data-pb-style="W5VIMR7">
-                        <div class="pagebuilder-overlay pagebuilder-poster-overlay" data-overlay-color=""
-                            data-element="overlay" data-pb-style="TEBEEB2">
-                            <div class="pagebuilder-poster-content">
-                                <div data-element="content"></div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div data-content-type="slide" data-slide-name="" data-appearance="poster" data-show-button="never"
-                data-show-overlay="never" data-element="main">
-                <div data-element="empty_link">
-                    <div class="pagebuilder-slide-wrapper"
-                        data-background-images="{\&quot;desktop_image\&quot;:\&quot;{{media url=wysiwyg/1.jpg}}\&quot;,\&quot;mobile_image\&quot;:\&quot;{{media url=wysiwyg/1a.jpg}}\&quot;}"
-                        data-background-type="image" data-video-loop="true" data-video-play-only-visible="true"
-                        data-video-lazy-load="true" data-video-fallback-src="" data-element="wrapper"
-                        data-pb-style="X5PYI8J">
-                        <div class="pagebuilder-overlay pagebuilder-poster-overlay" data-overlay-color=""
-                            data-element="overlay" data-pb-style="GL4B38B">
-                            <div class="pagebuilder-poster-content">
-                                <div data-element="content"></div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div data-content-type="slide" data-slide-name="" data-appearance="poster" data-show-button="never"
-                data-show-overlay="never" data-element="main">
-                <div data-element="empty_link">
-                    <div class="pagebuilder-slide-wrapper"
-                        data-background-images="{\&quot;desktop_image\&quot;:\&quot;{{media url=wysiwyg/2.jpg}}\&quot;,\&quot;mobile_image\&quot;:\&quot;{{media url=wysiwyg/2a.jpg}}\&quot;}"
-                        data-background-type="image" data-video-loop="true" data-video-play-only-visible="true"
-                        data-video-lazy-load="true" data-video-fallback-src="" data-element="wrapper"
-                        data-pb-style="G1KKEN9">
-                        <div class="pagebuilder-overlay pagebuilder-poster-overlay" data-overlay-color=""
-                            data-element="overlay" data-pb-style="WXHSFDH">
-                            <div class="pagebuilder-poster-content">
-                                <div data-element="content"></div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div data-content-type="slide" data-slide-name="" data-appearance="poster" data-show-button="never"
-                data-show-overlay="never" data-element="main">
-                <div data-element="empty_link">
-                    <div class="pagebuilder-slide-wrapper"
-                        data-background-images="{\&quot;desktop_image\&quot;:\&quot;{{media url=wysiwyg/3.jpg}}\&quot;,\&quot;mobile_image\&quot;:\&quot;{{media url=wysiwyg/3a.jpg}}\&quot;}"
-                        data-background-type="image" data-video-loop="true" data-video-play-only-visible="true"
-                        data-video-lazy-load="true" data-video-fallback-src="" data-element="wrapper"
-                        data-pb-style="JMQW2EB">
-                        <div class="pagebuilder-overlay pagebuilder-poster-overlay" data-overlay-color=""
-                            data-element="overlay" data-pb-style="SSS7R2N">
-                            <div class="pagebuilder-poster-content">
-                                <div data-element="content"></div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
