--- PageBuilderBlockTheme/composer.json
@@ -1,17 +1,0 @@
-{
-    "name": "repoback/pagebuilderblocktheme",
-    "description": "",
-    "type": "magento2-module",
-    "version": "1.0.0",
-    "license": [
-      "MIT"
-    ],
-    "autoload": {
-      "psr-4": {
-        "repoback\\PageBuilderBlockTheme\\": ""
-      },
-      "files": [
-        "registration.php"
-      ]
-    }
-  }
