+++ composer.json
@@ -1,8 +1,8 @@
 {
-    "name": "repoback/holamundow",
+    "name": "bootcamp/repoback",
     "description": "",
     "require": {
-      "php": ">=7.4"
+      "php": "^7.0"
     },
     "type": "magento2-module",
     "version": "1.0.3",
@@ -11,7 +11,7 @@
     ],
     "autoload": {
       "psr-4": {
-        "Repoback\\HolaMundoW\\": ""
+        "Bootcamp\\Repoback\\": ""
       },
       "files": [
         "registration.php"
