+++ i18n/es_CO.csv
@@ -1,0 +1,3 @@
+"Hello World karen","Hola Mundo karen"
+"Hello World leonardo","Hola Mundo leonardo"
+"Hello World wilbert","Hola Mundo wilbert"
