+++ view/frontend/layout/helloworld_wilbert_index.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="1column" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
+    <body>
+        <referenceContainer name="content">
+            <block  name="bootcamp.holamundo.wilbert" template="Bootcamp_Repoback::wilbert.phtml" />
+        </referenceContainer>
+    </body>
+</page>
