+++ view/frontend/templates/index.phtml
@@ -1,0 +1,5 @@
+<?php
+/** @var \Magento\Framework\View\Element\Template $block */
+?>
+
+<h1><?= __("Hello World") ?></h1>
